fix page mix up and flow
